Add script to test Hugging Face requests
