+++ .env.example
@@ -1,2 +1,4 @@
 DISCORD_TOKEN=your-bot-token-here
 DISCORD_CHANNEL_ID=channel-id-to-send-jokes
+DISCORD_GUILD_ID=your-guild-id
+HF_TOKEN=your-huggingface-token
