+++ index.js
@@ -1,8 +1,7 @@
-// bot.js – roasts via Hugging Face Llama-3, auto + on-demand, logs fetch errors
+// Roast bot using Hugging Face, auto + on-demand, logs fetch errors
 const { Client, GatewayIntentBits } = require("discord.js");
-// Use built-in fetch on Node 18+, fall back to node-fetch for older versions
-const fetch =
-  global.fetch || ((...a) => import("node-fetch").then(({ default: f }) => f(...a)));
+// Fetch roast replies from Hugging Face
+const getRoast = require("./roast");
 require("dotenv").config();
 
 const {
@@ -28,64 +27,7 @@ const client = new Client({
 });
 
 // ---------- Hugging Face generation -----------------------------------
-const MODEL = "meta-llama/Meta-Llama-3-8B-Instruct";
-const failMessage =
-  "Sorry, but I seem to be having trouble accessing my AI brain. Ask the mods for help (not @ggoollpp, he will annoy you)";
-
-async function getRoast(name = "friend") {
-  const prompt = `Give me one short, savage but playful roast for a Discord user named "${name}".`;
-  const url = `https://api-inference.huggingface.co/models/${MODEL}`;
-  const bodyData = {
-    inputs: prompt,
-    parameters: { max_new_tokens: 32, temperature: 0.9 },
-  };
-  const opts = {
-    method: "POST",
-    headers: {
-      Authorization: `Bearer ${HF_TOKEN}`,
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(bodyData),
-  };
-  try {
-    const r = await fetch(url, opts);
-    const respText = await r.text();
-    if (!r.ok) {
-      console.error("[HF] Request failed", {
-        request: { url, method: "POST", body: bodyData },
-        response: {
-          status: r.status,
-          statusText: r.statusText,
-          body: respText,
-        },
-      });
-      return failMessage;
-    }
-    let j;
-    try {
-      j = JSON.parse(respText);
-    } catch (e) {
-      console.error("[HF] JSON parse error", {
-        request: { url, method: "POST", body: bodyData },
-        response: respText,
-      });
-      return failMessage;
-    }
-    const text = j[0]?.generated_text?.replace(prompt, "").trim();
-    if (text) return text;
-    console.error("[HF] Unexpected response", {
-      request: { url, method: "POST", body: bodyData },
-      response: j,
-    });
-    return failMessage;
-  } catch (err) {
-    console.error("[HF] Roast fetch error", {
-      request: { url, method: "POST", body: bodyData },
-      error: err.message,
-    });
-    return failMessage;
-  }
-}
+// getRoast is provided by ./roast
 
 // ---------- activity & cooldown --------------------------------------
 const lastSeen = new Map(); // userId → last message time
