+++ package.json
@@ -8,7 +8,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
+    "start": "node index.js",
+    "hf-test": "node test-fetch.js"
   },
   "dependencies": {
     "discord.js": "^14.11.0",
